Add files via upload
Autonomous robot which was capable to locate red beacon and turn it off. After turning off the red beacon it located the green led beacon, went to it and grabbed it to the starting position. It's in C language, mechanical design which consists of body, arm of body. Along with electronics was involved soldering and Infrared Receiver Board.